requirements.txt
django-dept-directory
